+++ angel_bridge/src/mypage/PageCompo.tsx
@@ -1,9 +1,10 @@
 'use client'
 import { useState } from 'react'
-import Button from './memberInfo/components/Button'
 import Info from './memberInfo/components/Info'
 import { button_container } from './memberInfo/styles/button.css'
 import Wrapper from './payment/components/Wrapper'
+import InfoBtn from './memberInfo/common/components/InfoBtn'
+import IsPaidBtn from './memberInfo/common/components/IsPaidBtn'
 
 export default function PageCompo() {
   const [isInfo, setIsInfo] = useState(true)
@@ -19,8 +20,8 @@ export default function PageCompo() {
   return (
     <div>
       <div className={button_container}>
-        <Button text="회원정보" onClick={handleSwitchToInfo} />
-        <Button text="결제내역" onClick={handleSwitchToPaied} />
+        <InfoBtn isInfo={isInfo} onClick={handleSwitchToInfo} />
+        <IsPaidBtn isInfo={isInfo} onClick={handleSwitchToPaied} />
       </div>
       {isInfo ? <Info /> : <Wrapper />}
     </div>
