+++ angel_bridge/src/mypage/memberInfo/common/components/InfoBtn.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { ButtonProps } from '../types/buttonprops'
+import { buttonblack, buttongray } from '../styles/button.css'
+
+export default function InfoBtn({ onClick, isInfo }: ButtonProps) {
+  return (
+    <p onClick={onClick} className={isInfo ? buttonblack : buttongray}>
+      회원정보
+    </p>
+  )
+}
