+++ angel_bridge/src/mypage/memberInfo/common/styles/button.css.ts
@@ -1,0 +1,25 @@
+import { title24SB, vars } from '@/style/theme.css'
+import { style } from '@vanilla-extract/css'
+
+export const buttonblack = style([
+  title24SB,
+  {
+    color: vars.grayScale.grayscale800,
+  },
+])
+
+export const buttongray = style([
+  title24SB,
+  {
+    color: vars.grayScale.grayscale300,
+  },
+])
+
+export const button_container = style({
+  display: 'flex',
+  alignItems: 'center',
+  width: '100%',
+  gap: '2.4rem',
+
+  cursor: 'pointer',
+})
