+++ angel_bridge/src/mypage/memberInfo/common/types/buttonprops.ts
@@ -1,0 +1,4 @@
+export interface ButtonProps {
+  onClick: () => void
+  isInfo: boolean
+}
