+++ angel_bridge/src/mypage/memberInfo/styles/button.css.ts
@@ -1,19 +1,5 @@
-import { title20SB, title24SB, vars } from '@/style/theme.css'
+import { title20SB, vars } from '@/style/theme.css'
 import { style } from '@vanilla-extract/css'
-
-export const buttonblack = style([
-  title24SB,
-  {
-    color: vars.grayScale.grayscale800,
-  },
-])
-
-export const buttongray = style([
-  title24SB,
-  {
-    color: vars.grayScale.grayscale300,
-  },
-])
 
 export const button_container = style({
   display: 'flex',
