+++ angel_bridge/src/mypage/memberInfo/styles/info.css.ts
@@ -7,6 +7,7 @@ export const container = style({
   justifyContent: 'space-between',
   alignItems: 'center',
   width: '104.rem',
+  height: '57.2rem',
 })
 
 export const label = style([
