+++ angel_bridge/src/mypage/payment/components/Content.tsx
@@ -1,0 +1,23 @@
+import React from 'react'
+import Photo from './Photo'
+import Texts from './Texts'
+import CancelBtn from './CancelBtn'
+import {
+  container,
+  content,
+  photo_text_container,
+} from '../styles/container.css'
+
+export default function Content() {
+  return (
+    <div className={content}>
+      <div className={container}>
+        <div className={photo_text_container}>
+          <Photo />
+          <Texts />
+        </div>
+        <CancelBtn />
+      </div>
+    </div>
+  )
+}
